+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Configurable Agent API calls timeout.
+- New `NODE_NEW` event.
+- New `SHARD_ALLOCATION_NEW` event.
 
 ### Changed
 - **BREAKING**: Flatten encoded events structure (`data`, `event`, `timeout` are all root attributes).